Move non-GPL source files to separate folder
