+++ pynicotine/external/README.md
@@ -1,0 +1,38 @@
+# IP2Location
+
+## Country Database
+
+- Last updated on September 9, 2022
+- Original file name "IP2LOCATION-LITE-DB1.BIN" renamed to "ipcountrydb.bin"
+- Fetched from https://lite.ip2location.com/database/ip-country
+
+### License
+
+Attribution-ShareAlike 4.0 International (CC BY-SA 4.0)
+
+Copyright (c) 2001-2022 Hexasoft Development Sdn. Bhd.
+
+## Source Code
+
+Country database reader, modified for Nicotine+  
+Original source code: https://github.com/chrislim2888/IP2Location-Python
+
+### License
+
+MIT
+
+Copyright (c) 2021-2022 Nicotine+ Contributors  
+Copyright (c) 2017 IP2Location.com
+
+
+# tinytag
+
+Audio metadata parser, modified for Nicotine+  
+Original source code: https://github.com/devsnd/tinytag
+
+## License
+
+MIT
+
+Copyright (c) 2020-2022 Nicotine+ Contributors  
+Copyright (c) 2014-2022 Tom Wallroth
