+++ pynicotine/geoip.py
@@ -276,7 +276,7 @@ class GeoIP:
 
     def __init__(self):
 
-        from pynicotine.geoip.ip2location import IP2Location
+        from pynicotine.external.ip2location import IP2Location
         self.ip2location = IP2Location(os.path.join(os.path.dirname(__file__), "ipcountrydb.bin"))
 
     @classmethod
